dwarf_loader: Add parameters list to inlined expansion
The parameters of inlined expansions are stored in the lexblock tag
list, making it difficult to iterate over the parameters of a given
inlined expansion.

Add a list of parameters to 'struct inline_expansion', so the parameters
are stored in their expansion instead of the lexblock.

Omitted the storage of struct location in parameters here as a later
patch will persist location information until after CUs have gone;
this is necessary to ensure we can delay adding it until base BTF
has been generated.

Signed-off-by: Thierry Treyer <ttreyer@meta.com>
Signed-off-by: Alan Maguire <alan.maguire@oracle.com>